refactor(imports): changes to plugins dir structure
no nested plugin dir; all plugins reexported from plugins/index
