+++ js/index.d.ts
@@ -8,7 +8,4 @@ export * from 'jolocom-lib';
 /**
  * Re-exporting available plugins
  */
-export { DeepLinkingProvider } from './plugins/deep-linking/DeepLinkingProvider';
-export { JolocomDeepLinkingTransport } from './plugins/deep-linking/JolocomDeepLinkingTransport';
-export { LinkingErrorCode, JolocomLinking, } from './plugins/linking/JolocomLinking';
-export { JolocomWebSockets } from './plugins/web-sockets/JolocomWebSockets';
+export { DeepLinkingProvider, JolocomDeepLinkingTransport, LinkingErrorCode, JolocomLinking, JolocomWebSockets } from './plugins';
