+++ js/plugins/JolocomLinking.d.ts
