+++ js/plugins/JolocomLinking.js
