+++ js/plugins/JolocomWebSockets.d.ts
