+++ js/plugins/JolocomWebSockets.js
