--- js/plugins/deep-linking/DeepLinkingProvider.d.ts
@@ -1,3 +1,0 @@
-export declare abstract class DeepLinkingProvider {
-    abstract subscribe(handler?: (token: string) => Promise<void>): void;
-}
