--- js/plugins/deep-linking/DeepLinkingProvider.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.DeepLinkingProvider = void 0;
-class DeepLinkingProvider {
-}
-exports.DeepLinkingProvider = DeepLinkingProvider;
