+++ js/plugins/index.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var JolocomDeepLinkingTransport_1 = require("./JolocomDeepLinkingTransport");
+Object.defineProperty(exports, "JolocomDeepLinkingTransport", { enumerable: true, get: function () { return JolocomDeepLinkingTransport_1.JolocomDeepLinkingTransport; } });
+Object.defineProperty(exports, "DeepLinkingProvider", { enumerable: true, get: function () { return JolocomDeepLinkingTransport_1.DeepLinkingProvider; } });
+var JolocomLinking_1 = require("./JolocomLinking");
+Object.defineProperty(exports, "JolocomLinking", { enumerable: true, get: function () { return JolocomLinking_1.JolocomLinking; } });
+Object.defineProperty(exports, "LinkingErrorCode", { enumerable: true, get: function () { return JolocomLinking_1.LinkingErrorCode; } });
+var JolocomWebSockets_1 = require("./JolocomWebSockets");
+Object.defineProperty(exports, "JolocomWebSockets", { enumerable: true, get: function () { return JolocomWebSockets_1.JolocomWebSockets; } });
