+++ ts/index.ts
@@ -12,13 +12,13 @@ export * from 'jolocom-lib'
 /**
  * Re-exporting available plugins
  */
-export { DeepLinkingProvider } from './plugins/deep-linking/DeepLinkingProvider'
-export { JolocomDeepLinkingTransport } from './plugins/deep-linking/JolocomDeepLinkingTransport'
 export {
+  DeepLinkingProvider,
+  JolocomDeepLinkingTransport,
   LinkingErrorCode,
   JolocomLinking,
-} from './plugins/linking/JolocomLinking'
-export { JolocomWebSockets } from './plugins/web-sockets/JolocomWebSockets'
+  JolocomWebSockets
+} from './plugins'
 
 /**
  * Some hax required for Reach Native
