+++ ts/plugins/JolocomLinking.ts
