+++ ts/plugins/JolocomWebSockets.ts
