--- ts/plugins/deep-linking/DeepLinkingProvider.ts
@@ -1,3 +1,0 @@
-export abstract class DeepLinkingProvider {
-  public abstract subscribe(handler?: (token: string) => Promise<void>): void
-}
