+++ ts/plugins/index.ts
@@ -1,0 +1,3 @@
+export { JolocomDeepLinkingTransport, DeepLinkingProvider } from './JolocomDeepLinkingTransport'
+export { JolocomLinking, LinkingErrorCode } from './JolocomLinking'
+export { JolocomWebSockets } from './JolocomWebSockets'
